requirements.txt
fastapi>=0.115.5
uvicorn>=0.32.1
pydantic>=2.6.1,<3.0.0
google-generativeai>=0.3.0
yfinance>=0.2.36
agentql==1.8.1
playwright>=1.49.0
google-cloud-language==2.13.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.117.0
python-dotenv==1.0.1
aiohttp>=3.9.1
weave==0.51.33
sse-starlette==1.6.5
aperturedb[complete,server]>=0.4.40
agno>=1.1.3
hyperpocket>=0.3.6
hyperpocket_gemini>=0.1.0
youtube_transcript_api>=0.6.2
